refactor: 컨슈머 클래스 분리
